# Conflicts:
#	server/home/home-service/src/main/java/io/holoinsight/server/home/biz/listener/CustomPluginUpdateListener.java

+++ server/home/home-web/src/main/java/io/holoinsight/server/home/web/controller/AlarmDingDingRobotFacadeImpl.java
@@ -3,23 +3,22 @@
  */
 package io.holoinsight.server.home.web.controller;
 
-import io.holoinsight.server.home.biz.service.AlertDingDingRobotService;
-import io.holoinsight.server.home.biz.service.UserOpLogService;
-import io.holoinsight.server.home.common.util.MonitorException;
-import io.holoinsight.server.home.common.util.scope.AuthTargetType;
-import io.holoinsight.server.home.common.util.scope.MonitorScope;
-import io.holoinsight.server.home.common.util.scope.MonitorUser;
-import io.holoinsight.server.home.common.util.scope.PowerConstants;
-import io.holoinsight.server.home.common.util.scope.RequestContext;
-import io.holoinsight.server.home.dal.model.OpType;
-import io.holoinsight.server.home.dal.model.dto.AlarmDingDingRobotDTO;
-import io.holoinsight.server.home.facade.page.MonitorPageRequest;
-import io.holoinsight.server.home.facade.page.MonitorPageResult;
-import io.holoinsight.server.home.web.common.ManageCallback;
-import io.holoinsight.server.home.web.common.ParaCheckUtil;
-import io.holoinsight.server.home.web.interceptor.MonitorScopeAuth;
 import io.holoinsight.server.common.J;
 import io.holoinsight.server.common.JsonResult;
+import io.holoinsight.server.common.service.AlertDingDingRobotService;
+import io.holoinsight.server.common.service.UserOpLogService;
+import io.holoinsight.server.common.scope.AuthTargetType;
+import io.holoinsight.server.common.scope.MonitorScope;
+import io.holoinsight.server.common.scope.MonitorUser;
+import io.holoinsight.server.common.scope.PowerConstants;
+import io.holoinsight.server.common.RequestContext;
+import io.holoinsight.server.home.dal.model.OpType;
+import io.holoinsight.server.common.dao.entity.dto.AlarmDingDingRobotDTO;
+import io.holoinsight.server.common.MonitorPageRequest;
+import io.holoinsight.server.common.MonitorPageResult;
+import io.holoinsight.server.common.ManageCallback;
+import io.holoinsight.server.home.web.interceptor.MonitorScopeAuth;
+import io.holoinsight.server.home.web.security.LevelAuthorizationAccess;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -41,7 +40,8 @@ import java.util.Date;
 @RequestMapping("/webapi/alarmDingDingRobot")
 public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
 
-  private static String dingdingUrlPrefix = "https://oapi.dingtalk.com/robot/send?access_token=";
+  public static final String dingdingUrlPrefix =
+      "https://oapi.dingtalk.com/robot/send?access_token=";
 
   @Autowired
   private AlertDingDingRobotService alarmDingDingRobotService;
@@ -49,18 +49,16 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
   @Autowired
   private UserOpLogService userOpLogService;
 
+  @LevelAuthorizationAccess(paramConfigs = {"PARAMETER" + ":$!alarmDingDingRobotDTO"},
+      levelAuthorizationCheckeClass = "io.holoinsight.server.home.web.security.custom.AlarmDingDingRobotFacadeImplChecker")
   @PostMapping("/create")
   @ResponseBody
   @MonitorScopeAuth(targetType = AuthTargetType.TENANT, needPower = PowerConstants.EDIT)
-  public JsonResult<Long> save(@RequestBody AlarmDingDingRobotDTO alarmDingDingRobotDTO) {
+  public JsonResult<Long> create(@RequestBody AlarmDingDingRobotDTO alarmDingDingRobotDTO) {
     final JsonResult<Long> result = new JsonResult<>();
     facadeTemplate.manage(result, new ManageCallback() {
       @Override
-      public void checkParameter() {
-        ParaCheckUtil.checkParaNotBlank(alarmDingDingRobotDTO.getGroupName(), "groupName");
-        ParaCheckUtil.checkParaStartWith(alarmDingDingRobotDTO.getRobotUrl(), dingdingUrlPrefix,
-            "robotUrl");
-      }
+      public void checkParameter() {}
 
       @Override
       public void doManage() {
@@ -69,9 +67,11 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
         if (null != mu) {
           alarmDingDingRobotDTO.setCreator(mu.getLoginName());
         }
+        alarmDingDingRobotDTO.setModifier(null);
         if (null != ms && !StringUtils.isEmpty(ms.tenant)) {
           alarmDingDingRobotDTO.setTenant(ms.tenant);
         }
+        alarmDingDingRobotDTO.setWorkspace(requestContextAdapter.getWorkspace(true));
         alarmDingDingRobotDTO.setGmtCreate(new Date());
         alarmDingDingRobotDTO.setGmtModified(new Date());
         Long id = alarmDingDingRobotService.save(alarmDingDingRobotDTO);
@@ -87,6 +87,8 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
     return result;
   }
 
+  @LevelAuthorizationAccess(paramConfigs = {"PARAMETER" + ":$!alarmDingDingRobotDTO"},
+      levelAuthorizationCheckeClass = "io.holoinsight.server.home.web.security.custom.AlarmDingDingRobotFacadeImplChecker")
   @PostMapping("/update")
   @ResponseBody
   @MonitorScopeAuth(targetType = AuthTargetType.TENANT, needPower = PowerConstants.EDIT)
@@ -94,36 +96,20 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
     final JsonResult<Boolean> result = new JsonResult<>();
     facadeTemplate.manage(result, new ManageCallback() {
       @Override
-      public void checkParameter() {
-        ParaCheckUtil.checkParaNotNull(alarmDingDingRobotDTO.getId(), "id");
-        ParaCheckUtil.checkParaNotNull(alarmDingDingRobotDTO.getTenant(), "tenant");
-        ParaCheckUtil.checkParaStartWith(alarmDingDingRobotDTO.getRobotUrl(), dingdingUrlPrefix,
-            "robotUrl");
-        ParaCheckUtil.checkEquals(alarmDingDingRobotDTO.getTenant(),
-            RequestContext.getContext().ms.getTenant(), "tenant is illegal");
-
-      }
+      public void checkParameter() {}
 
       @Override
       public void doManage() {
 
         AlarmDingDingRobotDTO item = alarmDingDingRobotService
             .queryById(alarmDingDingRobotDTO.getId(), RequestContext.getContext().ms.getTenant());
-        if (null == item) {
-          throw new MonitorException("cannot find record: " + alarmDingDingRobotDTO.getId());
-        }
-        if (!item.getTenant().equalsIgnoreCase(alarmDingDingRobotDTO.getTenant())) {
-          throw new MonitorException("the tenant parameter is invalid");
-        }
 
         MonitorUser mu = RequestContext.getContext().mu;
         if (null != mu) {
           alarmDingDingRobotDTO.setModifier(mu.getLoginName());
         }
         MonitorScope ms = RequestContext.getContext().ms;
-        if (null != ms && !StringUtils.isEmpty(ms.tenant)) {
-          alarmDingDingRobotDTO.setTenant(ms.tenant);
-        }
+
         alarmDingDingRobotDTO.setGmtModified(new Date());
         boolean save = alarmDingDingRobotService.updateById(alarmDingDingRobotDTO);
 
@@ -138,16 +124,16 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
     return result;
   }
 
+  @LevelAuthorizationAccess(paramConfigs = {"PARAMETER" + ":$!id"},
+      levelAuthorizationCheckeClass = "io.holoinsight.server.home.web.security.custom.AlarmDingDingRobotFacadeImplChecker")
   @GetMapping("/query/{id}")
   @ResponseBody
-  @MonitorScopeAuth(targetType = AuthTargetType.TENANT, needPower = PowerConstants.EDIT)
+  @MonitorScopeAuth(targetType = AuthTargetType.TENANT, needPower = PowerConstants.VIEW)
   public JsonResult<AlarmDingDingRobotDTO> queryById(@PathVariable("id") Long id) {
     final JsonResult<AlarmDingDingRobotDTO> result = new JsonResult<>();
     facadeTemplate.manage(result, new ManageCallback() {
       @Override
-      public void checkParameter() {
-        ParaCheckUtil.checkParaNotNull(id, "id");
-      }
+      public void checkParameter() {}
 
       @Override
       public void doManage() {
@@ -161,15 +147,15 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
     return result;
   }
 
+  @LevelAuthorizationAccess(paramConfigs = {"PARAMETER" + ":$!id"},
+      levelAuthorizationCheckeClass = "io.holoinsight.server.home.web.security.custom.AlarmDingDingRobotFacadeImplChecker")
   @DeleteMapping(value = "/delete/{id}")
   @MonitorScopeAuth(targetType = AuthTargetType.TENANT, needPower = PowerConstants.EDIT)
   public JsonResult<Boolean> deleteById(@PathVariable("id") Long id) {
     final JsonResult<Boolean> result = new JsonResult<>();
     facadeTemplate.manage(result, new ManageCallback() {
       @Override
-      public void checkParameter() {
-        ParaCheckUtil.checkParaNotNull(id, "id");
-      }
+      public void checkParameter() {}
 
       @Override
       public void doManage() {
@@ -177,7 +163,8 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
         boolean rtn = false;
         AlarmDingDingRobotDTO alarmDingDingRobot =
             alarmDingDingRobotService.queryById(id, ms.getTenant());
-        if (alarmDingDingRobot != null) {
+        if (alarmDingDingRobot != null
+            && StringUtils.equals(alarmDingDingRobot.getTenant(), ms.getTenant())) {
           rtn = alarmDingDingRobotService.removeById(id);
         }
 
@@ -191,6 +178,8 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
     return result;
   }
 
+  @LevelAuthorizationAccess(paramConfigs = {"PARAMETER" + ":$!pageRequest"},
+      levelAuthorizationCheckeClass = "io.holoinsight.server.home.web.security.custom.AlarmDingDingRobotFacadeImplChecker")
   @PostMapping("/pageQuery")
   @ResponseBody
   @MonitorScopeAuth(targetType = AuthTargetType.TENANT, needPower = PowerConstants.VIEW)
@@ -199,9 +188,7 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
     final JsonResult<MonitorPageResult<AlarmDingDingRobotDTO>> result = new JsonResult<>();
     facadeTemplate.manage(result, new ManageCallback() {
       @Override
-      public void checkParameter() {
-        ParaCheckUtil.checkParaNotNull(pageRequest.getTarget(), "target");
-      }
+      public void checkParameter() {}
 
       @Override
       public void doManage() {
@@ -209,6 +196,7 @@ public class AlarmDingDingRobotFacadeImpl extends BaseFacade {
         if (null != ms && !StringUtils.isEmpty(ms.tenant)) {
           pageRequest.getTarget().setTenant(ms.tenant);
         }
+        pageRequest.getTarget().setWorkspace(requestContextAdapter.getWorkspace(true));
         JsonResult.createSuccessResult(result,
             alarmDingDingRobotService.getListByPage(pageRequest));
       }
